pyproject.toml
[tool.poetry]
name = "fastq-agent-system"
version = "0.4.0"
description = "Multi-agent system for FASTQ file quality analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "fastq_agents", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
anthropic = "^0.40.0"
biopython = "^1.84"
pydantic = "^2.0"
typer = "^0.12.0"
requests = "^2.31.0"
rich = "^13.0"
pandas = "^2.0"
matplotlib = "^3.8"
seaborn = "^0.13"
urllib3 = "^1.26.18"
python-dotenv = "^1.1.0"
plotly = "^6.2.0"
jinja2 = "^3.1.6"
weasyprint = "^65.1"
tomli = {markers = "python_version < \"3.11\"", version = "^2.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
isort = "^5.13"
flake8 = "^7.0"
mypy = "^1.8"
pre-commit = "^3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fastq-analyze = "fastq_agents.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
